Refactor metdata and add simple metdata caching (#81)
* Adding validation that cropmask is binary

* added (optional) savgol smoothing of LAI data to be more robust against noise from e.g missed clouds

* Adding smoothing options to config

* wip

* WIP cleaned stac downloader

* supporting multiple APSIM files in setup helper notebook

* WIP - setup for multiyear reporting

* Fixing stac downloader and adding more robust error handling

* Linting and starting to update the documentation

* updated documentation. Adding retries for download of imagery through STAC

* Removal old deprecated files

* Continuing to update documentation and adding quick logo :)

* minor update setup helper

* setup helper example updated

* WIP refactoring met data fetching: Seperate Scripts for NasaPower and ERA5. Using global caching for nasapower to avoid rate limiting. Also only merging Chirps data and other met data during the .met file creation

* Adding new cache-based nasapower fetching to pipeline

* fixing merge remainders

* Fix unfinished merge

* Fixing caching of nasapower and era5 data

* Adding restrictions on number of parallel match_sim_real_quicklook jobs - trying to avoid hitting max nproc on shared cluster resources due to high thread parallelism in the kaleido dependancy during image export

* removing old lai merge file

* ensuring different cache for mean and centroid based metdata

* Fixed displayig correct precipitation source in reporting. Updating documentation

* Updated Output documentation
