+++ requirements.txt
@@ -19,6 +19,7 @@ pytest
 pyarrow
 rasterio[s3]
 requests
+ruamel.yaml
 scipy>=1.14  # For Akima interpolator
 scikit-learn
 snakemake  # If using mamba/conda, install from the bioconda channel
