+++ vercye_ops/apsim/generate_weather_report.py
@@ -1,9 +1,12 @@
 """CLI to generate a weather report from CSV of weather"""
 
+from pathlib import Path
 import click
 import pandas as pd
 import plotly.graph_objs as go
 from plotly.subplots import make_subplots
+import pyarrow.parquet as pq
+
 
 MET_FILE_DTYPES = {
     'year': int,
@@ -23,14 +26,14 @@ MET_FILE_DTYPES = {
 @click.option('--precipitation_source', required=True, type=click.Choice(['NASA_POWER', 'CHIRPS', 'ERA5']), help='Source of precipitation data. "NASA_Power" or "CHIRPS".')
 @click.option('--precipitation_agg', required=True, type=click.Choice(['centroid', 'mean']), help='Aggregation method for precipitation data. "centroid" or "mean".')
 @click.option('--fallback_precipitation', default=False, help='Boolean specifying whether to fallback on nasapower/era5 precipitation data if chirps data is not available.')
-@click.option('--met_fpath', required=False, type=click.Path(exists=True), default=None, help='Path to the met CSV file.')
+@click.option('--chirps_file', required=False, type=click.Path(exists=True), default=None, help='Path to the chirps parquet file if using chirps.')
 @click.option('--header_lines', default=8, show_default=True, help='Number of lines in the header.')
 @click.option('--column_line', default=6, show_default=True, help='Line number of the column headers (1-based index).')
-def cli(input_fpath, output_fpath, precipitation_source, precipitation_agg, fallback_precipitation, met_fpath, header_lines, column_line):
+def cli(input_fpath, output_fpath, precipitation_source, precipitation_agg, fallback_precipitation, chirps_file, header_lines, column_line):
     """
     CLI wrapper to plot weather data from a .met file and save as an interactive HTML plot.
     """
-    data, metadata = plot_weather_data(input_fpath, precipitation_source, precipitation_agg, fallback_precipitation, met_fpath, header_lines, column_line)
+    data, metadata = plot_weather_data(input_fpath, precipitation_source, precipitation_agg, fallback_precipitation, chirps_file, header_lines, column_line)
     
     metadata.update({'input_fpath': input_fpath,
                      'output_fpath': output_fpath})
@@ -38,7 +41,7 @@ def cli(input_fpath, output_fpath, precipitation_source, precipitation_agg, fall
     fig.write_html(output_fpath)
 
 
-def plot_weather_data(file_path, precipitation_source, precipitation_agg, fallback_precipitation, met_fpath, header_lines=8, column_line=6):
+def plot_weather_data(file_path, precipitation_source, precipitation_agg, fallback_precipitation, chirps_fpath=None, header_lines=8, column_line=6):
     """
     Plot weather data from a .met file and save as an interactive HTML plot.
 
@@ -50,10 +53,10 @@ def plot_weather_data(file_path, precipitation_source, precipitation_agg, fallba
         Source of precipitation data. "NASA_POWER" or "CHIRPS".
     precipitation_agg : str
         Aggregation method for precipitation data. "centroid" or "mean".
-    met_fpath : str
-        Path to the met CSV file.
     fallback_precipitation: bool
         Flag if original precipitation is used if CHIRPS not available for ROI
+    chirps_fpath : str
+        Optional: Path to the chirps parquet file if chirps is being used.
     header_lines : int
         Number of lines in the header.
     column_line : int
@@ -96,20 +99,20 @@ def plot_weather_data(file_path, precipitation_source, precipitation_agg, fallba
     metadata['precipitation_source'] = precipitation_source
     metadata['precipitation_agg'] = precipitation_agg
 
-    # If precipitation source is CHIRPS, add original data for comparison
-    if precipitation_source.lower() == 'chirps':
-        df_met = pd.read_csv(met_fpath)
-
-        # Check if the Met CSV file contains any CHIRPS data, if not fallback was used
-        if not 'PRECTOTCORR_CHIRPS' in df_met.columns:
-            if not fallback_precipitation:
-                raise ValueError('No Chirps data found. Met CSV file does not contain the required column "PRECTOTCORR_CHIRPS.')
-            metadata['precipitation_source'] = f'Precipitatio Fallback.'
-            metadata['precipitation_agg'] = 'centroid'
-        else:
-            # Add unused original precipitation data for comparisons
-            df['original_precipitation_data'] = df_met['PRECTOTCORR']
+    # CHIRPS is not available globally - need to identify if CHIRPS or fallback was used
+    regions_using_chirps = []
+    if chirps_fpath is not None:
+        parquet_file = pq.ParquetFile(chirps_fpath)
+        regions_using_chirps = parquet_file.schema.names
+        region_name = "_".join(Path(file_path).stem.split('_')[:-1])  # Extract region name from file path
+        if region_name in regions_using_chirps:
             metadata['precipitation_source'] = 'CHIRPS'
+            metadata['precipitation_agg'] = precipitation_agg
+        else:
+            if not fallback_precipitation:
+                raise ValueError(f'No CHIRPS data found for region {region_name}. Please provide a valid region or set fallback_precipitation to True.')
+            metadata['precipitation_source'] = 'Precipitation Fallback (Met data)'
+            metadata['precipitation_agg'] = 'Met data aggregation method'
             
     
     # Extract date information
@@ -167,14 +170,7 @@ def create_plots(df, metadata):
                                  line=dict(color='goldenrod', dash=line_dash, width=1.5)), row=2, col=1)
 
         # Precipitation
-        if 'original_precipitation_data' in df.columns:
-            fig.add_trace(go.Scatter(x=df['date'], y=df['original_precipitation_data'], mode='lines', name=f'Original Precipitation (Unused)',
-                                     line=dict(color='darkgreen', dash=line_dash, width=1.5)), row=3, col=1)
-
-            fig.add_trace(go.Scatter(x=df['date'], y=df['rain'], mode='lines', name=f'Chirps Precipitation ({data_type})',
-                                    line=dict(color='dodgerblue', dash=line_dash, width=1.5)), row=3, col=1)
-        else:
-            fig.add_trace(go.Scatter(x=df['date'], y=df['rain'], mode='lines', name=f'Precipitation (Corrected) ({data_type})',
+        fig.add_trace(go.Scatter(x=df['date'], y=df['rain'], mode='lines', name=f'Precipitation (Corrected) ({data_type})',
                                     line=dict(color='navy', dash=line_dash, width=1.5)), row=3, col=1)
 
         # Wind
