+++ vercye_ops/lai/lai_creation_GEE/2_build_daily_vrts.py
@@ -11,8 +11,8 @@ import rasterio as rio
 
 
 def is_within_date_range(vf, start_date, end_date):
-    # files are expected to have the pattern f"{s2_dir}/{region}_{resolution}m_{date}_LAI_tile.tif"
-    date = Path(vf).stem.split("_")[-3]
+    # files are expected to have the pattern f"{s2_dir}/{region}_{resolution}m_{date}_LAI.tif"
+    date = Path(vf).stem.split("_")[-2]
     date = datetime.strptime(date, "%Y-%m-%d")
     return start_date <= date <= end_date
 
@@ -74,7 +74,7 @@ def main(lai_dir, out_dir, resolution, region_out_prefix, start_date, end_date):
     crs = []
     resolutions = []
 
-    lai_files = sorted(glob(f"{lai_dir}/*_{resolution}m_*_LAI_tile.tif"))
+    lai_files = sorted(glob(f"{lai_dir}/*_{resolution}m_*_LAI.tif"))
     if start_date is not None and end_date is not None:
         lai_files = [f for f in lai_files if is_within_date_range(f, start_date, end_date)]
 
@@ -106,7 +106,7 @@ def main(lai_dir, out_dir, resolution, region_out_prefix, start_date, end_date):
     # Group files by date
     date_groups = defaultdict(list)
     for region_file in lai_files:
-        date = region_file.split("_")[-3]
+        date = region_file.split("_")[-2]
         date_groups[date].append(region_file)
 
     # Use the most frequent resolution
