+++ vercye_ops/matching_sim_real/aggregate_yield_estimates.py
@@ -4,13 +4,14 @@ from pathlib import Path
 import click
 import pandas as pd
 import geopandas as gpd
+import pyarrow.parquet as pq
 
 from vercye_ops.utils.init_logger import get_logger
 
 logger = get_logger()
 
 
-def aggregate_yields(yield_dir, columns_to_keep):
+def aggregate_yields(yield_dir, columns_to_keep, chirps_path=None):
     """
     Aggregate yield estimates from multiple regions within a directory.
 
@@ -27,13 +28,17 @@ def aggregate_yields(yield_dir, columns_to_keep):
         A DataFrame containing aggregated yield information for all regions.
     """
     all_yields = []
+
+    regions_using_chirps = []
+    if chirps_path is not None:
+        parquet_file = pq.ParquetFile(chirps_path)
+        regions_using_chirps = parquet_file.schema.names
     
     for region_dir in Path(yield_dir).iterdir():
         if region_dir.is_dir():
             region_name = region_dir.name
             yield_estimate_csv_path = region_dir / f"{region_name}_converted_map_yield_estimate.csv"
             conv_factor_csv_path = region_dir / f"{region_name}_conversion_factor.csv"
-            met_csv_path = region_dir /  f"{region_name}_met.csv"
             lai_stats_csv_path = region_dir / f"{region_name}_LAI_STATS.csv"
             geojson_path = region_dir / f"{region_name}.geojson"
             
@@ -84,11 +89,7 @@ def aggregate_yields(yield_dir, columns_to_keep):
             combined_df = pd.merge(yield_df, conv_df, on='region', how='outer')
 
             # Add additional information for easier analysis
-            precipitation_src = None
-            if met_csv_path.exists():
-                df_met = pd.read_csv(met_csv_path)
-                # We expect the files to have been generated by the pipeline: If the 'PRECTOTCORR_CHIRPS' column exists, CHIRPS is used for simulation.
-                precipitation_src = 'CHIRPS' if 'PRECTOTCORR_CHIRPS' in df_met else 'Met Source'
+            precipitation_src = 'CHIRPS' if region_name in regions_using_chirps else 'Met Source'
             
             n_days_with_rs_data_valid = None
             mean_cloud_snow_percentage = None
@@ -138,8 +139,9 @@ def aggregate_yields(yield_dir, columns_to_keep):
 @click.option('--yield_dir', required=True, type=click.Path(exists=True), help='Path to the yield directory (e.g., year, timepoint) containing region subdirectories.')
 @click.option('--output_csv', required=True, type=click.Path(), help='Path to save the aggregated yield estimates CSV.')
 @click.option('--columns_to_keep', required=False, default=None, type=str, help='Comma-separated list of geojsons-columns to keep in the output CSV. If not provided, no additional columns from the geojsons will be kept.')
+@click.option('--chirps_path', required=False, type=click.Path(exists=True), default=None, help='Path to the CHIRPS Parquet file for precipitation data. Used to identify if fallback or chirps was used for precipitation.')
 @click.option('--verbose', is_flag=True, help='Enable verbose logging.')
-def cli(yield_dir, output_csv, columns_to_keep, verbose):
+def cli(yield_dir, output_csv, columns_to_keep, chirps_path, verbose):
     """Aggregate yield estimates from multiple regions within a directory."""
     if verbose:
         logger.setLevel('INFO')
@@ -147,7 +149,7 @@ def cli(yield_dir, output_csv, columns_to_keep, verbose):
         logger.setLevel('WARNING')
 
     logger.info(f"Processing directory: {yield_dir}")
-    aggregated_yields = aggregate_yields(yield_dir, columns_to_keep)
+    aggregated_yields = aggregate_yields(yield_dir, columns_to_keep, chirps_path)
     
     if aggregated_yields is not None:
         aggregated_yields.to_csv(output_csv, index=False)
