+++ vercye_ops/snakemake/profiles/hpc/config.yaml
@@ -2,6 +2,7 @@ cores: 110
 resources:
   - nasa_power_calls=5
   - era5_calls=10
+  - kaleido_resources=5
 
 # Might want to add the apsim limitations and match_sim to this aswell
 # Need to experiment with these values
